Solve Day 4: Printing Department
Some good old matrix adjacency. I've done problems like this dozens of
times in JavaScript at this point, and have abstracted most of the
nitty gritty away into helper functions, but I am going fresh in Go.

I went with a slice of strings rather than nested slices, which I think
was a fairly straightforward approach. No way I know of to check all the
adjacent cells without exhaustive conditionals, but I am so used to them
at this point, that it was painless and error free to write.