A simpler (and working) approach to Day 5 part 2
At this point I don't remember why I got it into my head that I should
separate the starts and ends and sort them in their own slices. That
approach seems to work in many cases, but clearly not all of them, and I
can't really follow the logic of it...

This working approach keeps the ranges together and sorts them by their
start. Then each can be included, dropped, or partially included
depending on how it compares to the end of the previously included
range. It's basically my first approach but without the separating.

Side note: Why the heck am I not allowed to use slices.SortedFunc on a
[][2]int64?? Compiler error was no help at all and neither was Google.
Thankfully sort.Slice seems to work, but it's not my first choice.
Something to look into later...